backend/requirements.txt
# backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
metricx==1.0.0
torch>=2.0.0
transformers[torch]>=4.30.0
numpy>=1.24.0
datasets>=2.13.1
accelerate>=0.20.0
sentencepiece==0.1.99
sacrebleu
unbabel-comet